requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
mcp>=1.26.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
